# Conflicts:
#	Engine/src/AGE/Assets/Public/AssetManager.h
#	Engine/src/AGE/Core/Private/DeviceManager.cpp
#	Engine/src/AGE/Core/Public/DeviceManager.h
#	Engine/src/AGE/Events/Public/RendererEvent.h
#	Engine/src/AGE/Math/Public/Vector3.h
#	Engine/src/AGE/Physics/Private/Physics2D.cpp
#	Engine/src/AGE/Physics/Private/World2D.cpp
#	Engine/src/AGE/Physics/Public/Physics2D.h
#	Engine/src/AGE/Platform/OpenGL/Private/OpenGLTexture.cpp
#	Engine/src/AGE/Platform/OpenGL/Private/OpenGLVertexArray.cpp
#	Engine/src/AGE/Platform/OpenGL/Public/OpenGLBuffer.h
#	Engine/src/AGE/Platform/Windows/Public/WindowsWindow.h
#	Engine/src/AGE/Project/Public/Project.h
#	Engine/src/AGE/Render/Private/Font.cpp
#	Engine/src/AGE/Render/Public/RenderAPI.h
#	Engine/src/AGE/Render/Public/RenderBuffer.h
#	Engine/src/AGE/Scene/Public/Components.h
#	Engine/src/AGE/Structs/Public/DataStructures.h
#	Engine/src/AGE/TileMap/Private/TileMapImporter.cpp
#	Engine/src/AGE/TileMap/Public/TileMapImporter.h
#	Engine/src/AGE/UI/Public/UIStructs.h
#	Engine/src/AGE/Utils/Private/Serializers.cpp
#	Engine/src/AGE/Utils/Private/WindowsUtils.cpp
#	Engine/src/AGE/VisualScripting/Private/NodeEditorWindow.cpp

+++ Editor/src/Editor_ImGui/Private/InventoryDatabase.cpp
@@ -61,7 +61,7 @@ namespace AGE
             ImGui::TableHeadersRow();
             ImGui::TableNextColumn();
 
-            for (int i = 0; i < m_Database.size(); i++)
+            for (size_t i = 0; i < m_Database.size(); i++)
             {
                 ImGui::InputText(std::string("##ItemName" + std::to_string(i)).c_str(), &m_Database[i].Name);
                 ImGui::TableNextColumn();
